Cleanup theme files
#1504
