+++ packages/components/src/theme/ThemeModel.ts
@@ -1,0 +1,17 @@
+export type BaseThemeType = 'dark' | 'light';
+export type BaseThemeKey = `default-${BaseThemeType}`;
+
+export type ThemePreloadStyleContent = `:root{${string}`;
+
+export interface ThemePreloadData {
+  baseThemeKey?: BaseThemeKey;
+  themeKey: string;
+  preloadStyleContent?: ThemePreloadStyleContent;
+}
+
+export interface ThemeData {
+  baseThemeKey?: BaseThemeKey;
+  themeKey: string;
+  name: string;
+  styleContent: string;
+}
