+++ packages/components/src/theme/ThemeUtils.ts
@@ -11,7 +11,10 @@ export function getThemeKey(pluginRootPath: string, themeName: string): string {
   return `${pluginRootPath}_${themeName.toLowerCase().replace(/\s/g, '-')}`;
 }
 
-// eslint-disable-next-line import/prefer-default-export
+/**
+ * Preload minimal theme variables from the cache.
+ * @param themeCache The theme cache to preload from
+ */
 export function preloadTheme(themeCache: ThemeCache): void {
   const preloadData = themeCache.getPreloadData();
 
@@ -36,131 +39,3 @@ export function preloadTheme(themeCache: ThemeCache): void {
     },
   ]);
 }
-
-// export function getSelectedThemes({
-//   getBaseTheme,
-//   getCustomTheme,
-//   getPreloadData,
-// }: ThemeCache): ThemeData[] {
-//   const { baseThemeKey, themeKey } = getPreloadData() ?? {};
-
-//   const base =
-//     getBaseTheme(baseThemeKey ?? themeKey) ?? getBaseTheme('default-dark');
-//   const custom = getCustomTheme(themeKey);
-
-//   return [base, custom].filter((t): t is ThemeData => t != null);
-// }
-
-// export function setSelectedThemes(
-//   { getBaseTheme, getCustomTheme, setPreloadData }: ThemeCache,
-//   selectedThemeKey: string
-// ): void {
-//   console.log('[TESTING] setSelectedThemes', selectedThemeKey);
-
-//   const theme =
-//     getBaseTheme(selectedThemeKey) ?? getCustomTheme(selectedThemeKey);
-
-//   if (theme == null) {
-//     // TODO: maybe clear cache?
-//     return;
-//   }
-
-//   const { baseThemeKey, themeKey } = theme;
-
-//   const preloadStyleContent = calculatePreloadStyleContent();
-
-//   setPreloadData({
-//     baseThemeKey,
-//     themeKey,
-//     preloadStyleContent,
-//   });
-// }
-
-// export type ThemeUtilsEventType = 'change';
-
-// export class ThemeUtils {
-//   constructor(private cache: ThemeCache) {
-//     bindAllMethods(this);
-//   }
-
-//   readonly eventListeners: Map<ThemeUtilsEventType, Set<() => void>> =
-//     new Map();
-
-//   getSelectedThemes(): ThemeData[] {
-//     const { baseThemeKey, themeKey } = this.cache.getPreloadData() ?? {};
-
-//     const base =
-//       this.cache.getBaseTheme(baseThemeKey ?? themeKey) ??
-//       this.cache.getBaseTheme('default-dark');
-
-//     const custom = this.cache.getCustomTheme(themeKey);
-
-//     return [base, custom].filter((t): t is ThemeData => t != null);
-//   }
-
-//   preloadTheme(): void {
-//     const preloadData = this.cache.getPreloadData();
-
-//     if (preloadData?.preloadStyleContent == null) {
-//       return;
-//     }
-
-//     const style = document.createElement('style');
-//     style.innerHTML = preloadData.preloadStyleContent;
-//     document.head.appendChild(style);
-
-//     this.cache.registerBaseTheme({
-//       name: 'Default Dark',
-//       themeKey: 'default-dark',
-//       styleContent: darkTheme,
-//     });
-
-//     this.cache.registerBaseTheme({
-//       name: 'Default Light',
-//       themeKey: 'default-light',
-//       styleContent: lightTheme,
-//     });
-
-//     this.eventListeners.get('change')?.forEach(handler => handler());
-//   }
-
-//   registerEventListener(
-//     eventType: ThemeUtilsEventType,
-//     handler: () => void
-//   ): () => void {
-//     if (!this.eventListeners.has(eventType)) {
-//       this.eventListeners.set(eventType, new Set());
-//     }
-
-//     this.eventListeners.get(eventType)?.add(handler);
-
-//     return () => {
-//       this.eventListeners.get(eventType)?.delete(handler);
-//     };
-//   }
-
-//   setSelectedThemes(selectedThemeKey: string): void {
-//     console.log('[TESTING] setSelectedThemes', selectedThemeKey);
-
-//     const theme =
-//       this.cache.getBaseTheme(selectedThemeKey) ??
-//       this.cache.getCustomTheme(selectedThemeKey);
-
-//     if (theme == null) {
-//       // TODO: maybe clear the cache?
-//       return;
-//     }
-
-//     const { baseThemeKey, themeKey } = theme;
-
-//     const preloadStyleContent = calculatePreloadStyleContent();
-
-//     this.cache.setPreloadData({
-//       baseThemeKey,
-//       themeKey,
-//       preloadStyleContent,
-//     });
-
-//     this.eventListeners.get('change')?.forEach(handler => handler());
-//   }
-// }
