+++ packages/components/src/theme/index.ts
@@ -1,3 +1,4 @@
 export * from './ThemeCache';
 export * from './ThemeContext';
+export * from './ThemeModel';
 export * from './ThemeUtils';
